[feat] #11 merge BbangZipTextFieldType.kt -> BbangZipTextFieldInputState.kt
